use double to represent the abstract Real type
git-svn-id: https://svn.code.sf.net/p/plexil/code/trunk@1667 d2be67b0-814c-0410-949f-d215b6b29573

+++ src/apps/sample/types.cc
@@ -46,7 +46,7 @@ ctype decode##type (type x) \
 } 
 
 defSimpleType(Int, int)
-defSimpleType(Real, float)
+defSimpleType(Real, double)
 defSimpleType(Bool, bool)
